tagging_website/serverside/requirements.txt
fastapi==0.115.11
SQLAlchemy==2.0.38
tweepy==4.15.0
uvicorn==0.34.0
python-jose==3.4.0
setuptools==75.8.2
psycopg2-binary==2.9.6
gunicorn==20.1.0
pydantic==2.10.6
requests==2.29.0
pandas==2.2.3
python-dotenv==1.0.1
ftfy==6.3.1